requirements.txt
# ExplainGrade AI - Complete Dependencies
# Install with: pip install -r requirements.txt

# Core NLP & ML
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
sentence-transformers>=2.7.0
spacy>=3.7.0
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# FastAPI & Backend
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6

# Real-Time Processing (Kafka)
kafka-python>=2.0.2

# Batch Processing (Spark)
pyspark>=3.4.0

# Database & Caching
psycopg2-binary>=2.9.0
pymongo>=4.5.0
redis>=5.0.0
elasticsearch>=8.9.0

# API & HTTP
requests>=2.31.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
python-json-logger>=2.0.7
